Drop parse() and readFile() methods from base Parser; monkeypatch them in getParser()
