+++ compare_locales/parsers/base.py
@@ -6,8 +6,6 @@ import re
 import bisect
 import codecs
 from collections import Counter
-from compare_locales.keyedtuple import KeyedTuple
-from compare_locales.paths import File
 
 __constructors = []
 
@@ -325,15 +323,6 @@ class Parser:
             self.encoding = "utf-8"
         self.ctx = None
 
-    def readFile(self, file):
-        """Read contents from disk, with universal_newlines"""
-        if isinstance(file, File):
-            file = file.fullpath
-        # python 2 has binary input with universal newlines,
-        # python 3 doesn't. Let's split code paths
-        with open(file, encoding=self.encoding, errors="replace", newline=None) as f:
-            self.readUnicode(f.read())
-
     def readContents(self, contents):
         """Read contents and create parsing context.
 
@@ -344,9 +333,6 @@ class Parser:
 
     def readUnicode(self, contents):
         self.ctx = self.Context(contents)
-
-    def parse(self):
-        return KeyedTuple(self)
 
     def __iter__(self):
         return self.walk(only_localizable=True)
