+++ compare_locales/tests/test_parser.py
@@ -5,61 +5,23 @@
 import pkg_resources
 import shutil
 import tempfile
-import textwrap
 import unittest
+from os.path import join
 
-from compare_locales import parser, mozpath
+from compare_locales.parser import Parser, getParser, patchParser
 
 
-class TestParserContext(unittest.TestCase):
-    def test_linecol(self):
-        "Should return 1-based line and column numbers."
-        ctx = parser.Parser.Context(
-            """first line
-second line
-third line
-"""
-        )
-        self.assertEqual(ctx.linecol(0), (1, 1))
-        self.assertEqual(ctx.linecol(1), (1, 2))
-        self.assertEqual(ctx.linecol(len("first line")), (1, len("first line") + 1))
-        self.assertEqual(ctx.linecol(len("first line") + 1), (2, 1))
-        self.assertEqual(ctx.linecol(len(ctx.contents)), (4, 1))
-
+class TestEmptyParser(unittest.TestCase):
     def test_empty_parser(self):
-        p = parser.Parser()
+        p = patchParser(Parser())
         entities = p.parse()
         self.assertTupleEqual(entities, tuple())
-
-
-class TestOffsetComment(unittest.TestCase):
-    def test_offset(self):
-        ctx = parser.Parser.Context(
-            textwrap.dedent(
-                """\
-            #foo
-            #bar
-            # baz
-            """
-            )
-        )  # noqa
-        offset_comment = parser.OffsetComment(ctx, (0, len(ctx.contents)))
-        self.assertEqual(
-            offset_comment.val,
-            textwrap.dedent(
-                """\
-                foo
-                bar
-                 baz
-            """
-            ),
-        )
 
 
 class TestUniversalNewlines(unittest.TestCase):
     def setUp(self):
         """Create a parser for this test."""
-        self.parser = parser.Parser()
+        self.parser = patchParser(Parser())
         self.dir = tempfile.mkdtemp()
 
     def tearDown(self):
@@ -68,7 +30,7 @@ class TestUniversalNewlines(unittest.TestCase):
         shutil.rmtree(self.dir)
 
     def test_universal_newlines(self):
-        f = mozpath.join(self.dir, "file")
+        f = join(self.dir, "file")
         with open(f, "wb") as fh:
             fh.write(b"one\ntwo\rthree\r\n")
         self.parser.readFile(f)
@@ -90,10 +52,10 @@ class TestPlugins(unittest.TestCase):
         pkg_resources.working_set.__setstate__(self.old_working_set_state)
 
     def test_dummy_parser(self):
-        p = parser.getParser("some/weird/file.ext")
+        p = getParser("some/weird/file.ext")
         self.assertIsInstance(p, DummyParser)
 
 
-class DummyParser(parser.Parser):
+class DummyParser(Parser):
     def use(self, path):
         return path.endswith("weird/file.ext")
